shards: move to conda/_private
