+++ conda/_private/shards/cache.py
