+++ pyproject.toml
@@ -74,11 +74,11 @@ exclude_lines = [
 omit = [
   "*/site-packages/*",
   '*\site-packages\*',
-  "_conda/shards/typing.py",
+  "conda/_private/shards/typing.py",
   "conda/_vendor.py",
-  "conda/console.py",
   "conda/cli/activate.py",
   "conda/cli/main_package.py",
+  "conda/console.py",
   "conda/exports.py",
   "conda/gateways/connection/adapters/ftp.py",
   "conda/gateways/connection/adapters/s3.py",
@@ -97,7 +97,7 @@ sort = "Miss"
 relative_files = true
 
 [tool.hatch.build]
-include = ["conda", "_conda"]
+include = ["conda"]
 
 [tool.hatch.build.hooks.vcs]
 version-file = "conda/_version.py"
