+++ tests/shards/conftest.py
@@ -25,14 +25,14 @@ import msgpack
 import pytest
 import zstandard
 
-from _conda.shards import cache, shards, subset
+from conda._private.shards import cache, shards, subset
 from conda.base.context import context, reset_context
 from conda.models.channel import Channel, all_channel_urls
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterable, Iterator
 
-    from _conda.shards.typing import RepodataDict, ShardDict, ShardsIndexDict
+    from conda._private.shards.typing import RepodataDict, ShardDict, ShardsIndexDict
 
 # Test channel names
 CONDA_FORGE_WITH_SHARDS = "conda-forge"
